+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Perfil tab updated. Now google acount photo, name and email are showed in profile tab. Total amount of cocoms ended are showed in profile tab, a button to go to reports tab added. [#5] (https://github.com/SAngelCuadrosV/ac-flutter-poc/pull/5) 
+Author: [@RojasVSantiago](https://github.com/RojasVSantiago)
 - Firebase conected with the app. Data can be write, when a cocom is finished send a post with the information, and read by users registered in Identity Platform, report shows the information about all cocom ended by the user. Additional report can be modify (only for start hout and end hour) in report tab. [#4](https://github.com/SAngelCuadrosV/ac-flutter-poc/pull/4)
 Author: [@RojasVSantiago](https://github.com/RojasVSantiago)
 - General changes in RouteTab. Now can be added locations to the cocom pressing the floating button. Locations can be removed dismissing them, confirmation requiered. [#3](https://github.com/SAngelCuadrosV/ac-flutter-poc/pull/3)